Prioritize First Discovered Resource Set
When iterating over a search result, if it finds a resource set as part of the results,
the first one should jump the line and be the first result.

This is necessary since special characters (specifically a period) may push the resource
set lower than a purview custom connector entity or folder. In addition since we are
only using FILTER the search scores are all 1.0 and cannot be further ordered.

Adding a test case to make sure we don't regress.