Correct Search Behavior
When Purview has a limit set to 1000, search scores are set to 1 and resulting in sub-optimal
search result ordering. The custom connector generic is showing up first when other assets
like sql tables and resource sets are showing up second.

This is solved by changing the default limit to 100 instead of 1000 (or any value less than 1000)
and Microsoft Purview search scores are populating correctly.